changed module importing and opening
changed module importing and opening to support file opening & selecting files from multiple locations.

removed ModuleDir (not necessary to have everything there now)
added 3 button to add, remove, and open the loaded modules
the modules can be scattered on any folders now
marked a few locations with TODO that have to be completed

I still had some problems, when selecting start experiment, the code froze. Not sure if it's my issue or not.